### 2021-06-25 #3401 @whedon
[PRE REVIEW]: SHINE_color: controlling low-level properties of colorful images
**Submitting author:** @RodDalBen (<a href="http://orcid.org/0000-0003-2185-8762">Rodrigo Dal Ben</a>)
**Repository:** <a href="https://github.com/RodDalBen/SHINE_color" target ="_blank">https://github.com/RodDalBen/SHINE_color</a>
**Version:** 0.0.2
