### 2021-07-08 RodDalBen/SHINE_color#2 @da5nsy
'Statement of need' section required
As per [here](https://joss.readthedocs.io/en/latest/review_criteria.html#a-statement-of-need) and noted by @whedon [here](https://github.com/openjournals/joss-reviews/issues/3449#issuecomment-873371380).

~~At minimum, I think you could probably rename the summary section to 'Statement of need'.~~ I've just seen that [here](https://joss.readthedocs.io/en/latest/review_checklist.html#documentation) that both summary and statement of need sections are required.
