+++ paper/paper.md
@@ -23,7 +23,7 @@ One powerful way to control low-level properties of experimental stimuli is to u
 
 # Implementation
 
-The SHINE_color toolbox works in an intuitive way (\autoref{fig:Figure 1}). Once called in the command window of MATLAB, by typing SHINE_color, the script guides the user through a series of questions that specify the input files characteristics (either a set of images or a video) and the operations to be performed (luminance, histogram, Fourier amplitude spectra specification; \autoref{fig:Figure 1}). Following, the input RBG images are transformed to HSV images. If a video is provided, its’ frames are first extracted, then they are transformed to HSV color space. The HSV color space separates Hue, Saturation, and Value (luminance) channels. Once transformed, Hue and Saturation are hold in memory and are not manipulated, but the Value channel (originally ranging from 0 to 1) is rescaled to match greyscale range (from 0 to 255). Then, all operations from SHINE (Table 1) can be performed in the scaled Value channel. Following, the Value channel is rescaled to its’ original range (0-1) and is combined with its’ original Hue and Saturation channels. These HSV images are transformed back to RBG images. For videos, the frames are recombined back into a video. In addition, to quantify the changes, the mean and standard deviation of each image’s Value channel is automatically calculated before and after manipulations.
+The SHINE_color toolbox works in an intuitive way (\autoref{fig:Figure 1}). Once called in the command window of MATLAB, by typing SHINE_color, the script guides the user through a series of questions that specify the input files characteristics (either a set of images or a video) and the operations to be performed (luminance, histogram, Fourier amplitude spectra specification; \autoref{fig:Figure 1}). Following, the input RBG images are transformed to HSV images. If a video is provided, its’ frames are first extracted, then they are transformed to HSV color space. The HSV color space separates Hue, Saturation, and Value (luminance) channels. Once transformed, Hue and Saturation are held in memory and are not manipulated, but the Value channel (originally ranging from 0 to 1) is rescaled to match greyscale range (from 0 to 255). Then, all operations from SHINE (Table 1) can be performed in the scaled Value channel. Following, the Value channel is rescaled to its’ original range (0-1) and is combined with its’ original Hue and Saturation channels. These HSV images are transformed back to RBG images. For videos, the frames are recombined back into a video. In addition, to quantify the changes, the mean and standard deviation of each image’s Value channel is automatically calculated before and after manipulations.
 
 ![SHINE_color workflow. Green boxes are adaptations that allow SHINE operations to be performed on colorful images.\label{fig:Figure 1}](fig1.png)
 
