### 2021-07-08 #3449 @da5nsy
Noting this previous work: https://github.com/Ruedeerat/RGBshine
