### 2021-07-08 RodDalBen/SHINE_color#4 @da5nsy
Tutorial tweak
(as part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/3449))

In [the tutorial](https://github.com/RodDalBen/SHINE_color#tutorial), I think that either point 3 should direct the user to navigate to SHINE_color\toolbox, or there should be an additional point instructing the user to add the folder and subfolders to the path.
