### 2021-07-17 e2fed1f
update tutorial to reflect folder structure
