### 2021-07-17 2646d62
fix wording on tutorial
