### 2021-07-17 RodDalBen/SHINE_color#4 @RodDalBen
Done. I added a step and included instructions to go to the subfolder "SHINE_color/toolbox".
