### 2021-07-18 4d93f61
add summary and statement of need; add Ruedeerat citation; fix typos
