+++ paper/paper.bib
@@ -83,6 +83,13 @@ url = {http://doi.wiley.com/10.1111/desc.12281},
 volume = {18},
 year = {2015}
 }
+@article{Ruedeerat2018,
+author = {Ruedeerat},
+month = {feb},
+title = {{RGBshine}},
+url = {https://github.com/Ruedeerat/RGBshine},
+year = {2018}
+}
 @article{Sirois2014,
 abstract = {Pupillometry is the study of changes in the diameter of the pupil as a function of cognitive processing. This review paper provides a brief historical overview of the study of pupillometry in cognitive science. The physiology of pupillary responses is introduced, leading to an outline of early pupillometry work, which began with the seminal work of Hess and Polt in the 1960s. The paper then presents a broad review of contemporary research in cognitive sciences that relies on pupillometry. This review is organized around five general domains, namely perception, language processing, memory and decision making, emotion and cognition, and cognitive development. In order to illustrate the nature of the method, and the challenges of analysis, the next section of the review details the process of compiling, processing, and analyzing data from a simple, typical pupillometry study.},
 author = {Sirois, Sylvain and Brisson, Julie},
