### 2021-07-18 b4fcc50
add summary and statement of need; add Ruedeerat citation; fix typos
