+++ paper/paper.md
@@ -17,13 +17,17 @@ bibliography: paper.bib
 
 # Summary
 
+Pupil dilation responses are versatile measures that can be used to study an array of complex psychological phenomena. However, controlling stimuli properties is an essential step to minimize experimental confounds that can result in unwanted pupil dilation. Here we describe an adapted MATLAB toolbox for controlling properties of colorful images.
+
+# Statement of need
+
 Pupil dilation responses can be used to investigate an array of cognitive abilities across the lifespan [@Hepach2016; @Sirois2014; @Zhang2019; @Zhao2019a]. Whereas it is a versatile measure of high-level abilities, pupil dilation can be greatly affected by low-level properties of stimuli and experimental setting such as luminance of visual stimuli and experimental room [e.g., @Hepach2016; @Tsukahara2020].
 
 One powerful way to control low-level properties of experimental stimuli is to use the SHINE toolbox for MATLAB [@Willenbockel2010]. This toolbox contains a set of functions that allows users to precisely specify luminance and contrast, histogram, and Fourier amplitude spectra of visual stimuli. These parametric manipulations minimize potential low-level confounds when investigating higher-level processes (e.g., cognitive effort, recognition). However, SHINE only works with greyscale images. Whereas this serves well to many research purposes [e.g., @Lawson2017; @Rodger2015], other research goals might benefit from colorful images [e.g., @Cheng2019; @Hepach2016; @Zhang2019]. Here, we describe the SHINE_color, an adaptation of SHINE that allow users to perform all operations from SHINE toolbox to colorful images.
 
 # Implementation
 
-The SHINE_color toolbox works in an intuitive way (\autoref{fig:Figure 1}). Once called in the command window of MATLAB, by typing SHINE_color, the script guides the user through a series of questions that specify the input files characteristics (either a set of images or a video) and the operations to be performed (luminance, histogram, Fourier amplitude spectra specification; \autoref{fig:Figure 1}). Following, the input RBG images are transformed to HSV images. If a video is provided, its frames are first extracted, then they are transformed to HSV color space. The HSV color space separates Hue, Saturation, and Value (luminance) channels. Once transformed, Hue and Saturation are held in memory and are not manipulated, but the Value channel (originally ranging from 0 to 1) is rescaled to match greyscale range (from 0 to 255). Then, all operations from SHINE (Table 1) can be performed in the scaled Value channel. Following, the Value channel is rescaled to its original range (0-1) and is combined with its original Hue and Saturation channels. These HSV images are transformed back to RBG images. For videos, the frames are recombined back into a video. In addition, to quantify the changes, the mean and standard deviation of each image’s Value channel is automatically calculated before and after manipulations.
[…]
