### 2021-07-18 RodDalBen/SHINE_color#2 @RodDalBen
Right! I added a brief summary and renamed the former "summary" to Statement of need. Thank you!
