### 2021-07-18 #3449 @RodDalBen
Hi @da5nsy,
Following your review, I added the following in the new version of the manuscript:
- a new "Summary" section;
- the former "Summary" was renamed to "Statement of need";
- a citation of [RGBShine](https://github.com/Ruedeerat/RGBshine);
- fixed typos;
- accepted word changes;
- fixed tutorial pointing to the right subfolder `SHINE_color/toolbox`.
**Thank you** for the careful review!
