### 2021-07-18 RodDalBen/SHINE_color#6 @da5nsy
choice of HSV vs other color spaces?
(as part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/3449))

I have yet to dive into the code, but in the name of expediency, I think it appropriate to raise a question that I predict will arise:

**HSV is only loosely related to psychophysically or physiologically meaningful colorspaces. Is it the right tool to use here?**

More specifically, considering the range of potential ways in which one can convert from RGB to a (pseudo-)luminance channel, which choice was made? Do the results it produces suffice for pupilometry (and other visual psychophysics) experiments, and are the implicit assumptions appropriate?
Were alternatives such as CIELAB/CIELUV/DKL/[more recent color appearance spaces] considered?

- _["The issue with both HSV and HSL is that these approaches do not effectively separate colour into their three value components according to human perception of color"](https://en.wikipedia.org/wiki/HSL_and_HSV)_
- _["the definition of a lightness or value dimension is less obvious: there are several possibilities depending on the purpose and goals of the representation"](https://en.wikipedia.org/wiki/HSL_and_HSV#Lightness)_
- This [whole section](https://en.wikipedia.org/wiki/HSL_and_HSV#Disadvantages)

At a minimum, a discussion of how these issues are thought of is warranted in the paper.

_[Additional context: I know a fair amount about CIE colorimetry, but little about what I think of as 'computer scientist color-spaces' such as HSV]_
