### 2021-07-19 #3449 @RodDalBen
> We're only just getting started! 😉
>
> I've realised I need to read the original Willenbockel+ paper before I can comment further, but I'll come back to this once I've done so.

Nice, Willenbockel et al. is the mothership! Actually, the seed for the adaptation is on their paper (see p. 682, "Applying SHINE to Color Images").
