### 2021-07-19 RodDalBen/SHINE_color#6 @RodDalBen
This is a great question @da5nsy ! Let me give you the context for my decision for HSV.

My initial idea was to use the [DKL color space](https://alexholcombe.wordpress.com/2011/07/11/color-space-pictured-and-animated-derrington-krauskopf-lennie/), as the idea arose when working with Psychopy/Python, that has a built-in function [rgb2dkl](https://www.psychopy.org/general/colours.html). However, both SHINE and SHINE_color are powered by the [MATLAB Image Processing Toolbox](https://www.mathworks.com/help/images/understanding-color-spaces-and-color-space-conversion.html). This toolbox does not support DKL color space. There are custom [functions on the web](https://github.com/MCG-NKU/SalBenchmark/blob/master/Code/matlab/SS/rgb2dkl.m), but, at least for now, I prefer to stick with the toolbox.

My next choice was for [CIELab](https://en.wikipedia.org/wiki/CIELAB_color_space), which is supported by MATLAB toolbox. I tried it in a [related project](https://github.com/RodDalBen/lum_fun), but the resulting images were very odd and not so consistent when measured with a [luminance meter](https://sensing.konicaminolta.us/us/products/ls-160-luminance-meter/). I turned to HSV as an alternative color space to contrast it with CIELab results. HSV results were surprisingly good. The transformed images (from NOUN database, available as examples on the [INPUT folder from SHINE_color](https://github.com/RodDalBen/SHINE_color/tree/main/toolbox/SHINE_color_INPUT/samples)) were more appealing and the light meter measures were more consistent and closer to the ones with greyscale images from SHINE. Thus, with HSV, SHINE_color was doing what it is supposed to do: output images with similar physical lightness.

Now, I understand your concern. [Physical lightness is not always the same as perceived brightness](https://www.vis4.net/blog/2011/12/avoid-equidistant-hsv-colors/). So images with similar physical luminance could elicit different pupil dilations. My understanding is that researchers are usually very aware of this challenge. As I tried to point in the manuscript, controlling low-level properties of stimuli is just one step towards a valid pupillometry design. For instance, stimuli (that usually vary in complexity) must be empirically validated  with the target population (infants, children, adults will most likely respond differently to the same stimuli set) in the target cultural context. Also, inventive counterbalancing is necessary to minimize individual differences. On top of that, there are [several other "details"](https://doi.org/10.1016/j.cognition.2021.104643) to consider.

One way to give more flexibility to SHINE_color would be to have all color spaces supported by MATLAB Image Processing Toolbox (i.e., HSV, CIE XYZ, CIELab, YCbCr, YIQ). as options that the user could choose from (or, given enough memory, run them all at once and compare the results). What do you think about that?

Another alternative to have even more options would be to rely on color space transformations, like [this](https://www.mathworks.com/matlabcentral/fileexchange/28790-colorspace-transformations), [this](https://www.mathworks.com/matlabcentral/fileexchange/53354-colorspace-conversion-tools), plus the [rgb2dkl](https://github.com/MCG-NKU/SalBenchmark/blob/master/Code/matlab/SS/rgb2dkl.m), but I would rather stick with the Image Processing Toolbox for now.
