### 2021-09-02 RodDalBen/SHINE_color#7 @takuma929
JOSS review - comments
First of all sorry for the delay, but I provide summary and specific comments for JOSS review.

**Summary**
The author created a MATLAB toolbox to control low-level image features in the luminance channel of color images. This work directly builds on the SHINE toolbox which performs the same transformation for achromatic images.

I think that the toolbox can be useful for experiments that require careful control of low-level image statistics. Thanks for making this toolbox available to the public. I have only a few comments.

**Major comments**
(1) It would be good to consider including some example images in the paper to show how different the original color images and the manipulated color images appear, along with manipulated statistics (e.g. luminance histogram or power spectrum). Currently, it is not clear from the paper what the toolbox does without running the code or reading a previous paper (Willenbockel et al., 2010).

(2) line 41 - 43: “In addition, to quantify the changes, the mean and standard deviation of each image’s Value channel is automatically calculated before and after manipulations.”

As stated here, the code calculates the mean and standard deviation of the luminance channel and I found that these values seem to be saved to “img_statistics_pos.txt” after running the code. From a user’s point of view, I think it would be helpful to save figures of the manipulated statistics (e.g., luminance histogram) so that users can check whether the code did what it intended to do. Thus, please consider adding an option to save these images.

**Minor comments**
(1) The current summary section starts with “pupil dilation are versatile measures…”. I understand that the author may have a motivation to use this toolbox for studies involving measurement of pupil dilation response, but I feel the summary still should start by explaining what the toolbox does and then stating what it can be useful for. Thus, I would request that the author reorder the summary section.

(2) I think that we need at least two images to execute the scripts (e.g., histogram matching), but the current code does not give any error message even if I try to run it with only one image. Consider displaying an error message.
