### 2021-09-02 #3449 @takuma929
Hi @RodDalBen  sorry for the delay, but I have just posted my comments here.

https://github.com/RodDalBen/SHINE_color/issues/7#issue-986779868
