+++ toolbox/lum_calc.m
@@ -21,99 +21,128 @@
 %
 % Please send suggestions or corrections to dalbenwork@gmail.com
 % ------------------------------------------------------------------------
+% SHINE_color toolbox, September 2021, version 0.0.3
+% (c) Rodrigo Dal Ben
+%
+% - Fix formula for pooled SD;
+% - Add calculations based on colorspaces: 
+%   - channel 3 for hsv;
+%   - channel 1 for cielab;
+% - Add pre and pos luminance summary on a single file, saved in the new 
+% DIAGNOSTICS subfolder;
+% - Update progress msg;
+% - Update input for 'lum2scale' function
+% ------------------------------------------------------------------------
 
 
-function lum_calc (input_folder, output_folder, imformat)
+function lum_calc (input_folder, output_folder, imformat, cs)
 %% Initial setup
 % Set input and output source and name pattern
 src_input = dir(fullfile(strcat(input_folder, filesep, '*.', imformat))); 
 src_output = dir(fullfile(strcat(output_folder, filesep, '*.', imformat)));
+output_folder_diagnostics = fullfile(pwd,'SHINE_color_OUTPUT', 'DIAGNOSTICS');
 
 % Set number of im
 numim = length(src_input);
 
 % Opening the output .txt
-statistics_pre = fopen([input_folder filesep 'img_statistics_pre.txt'], 'wt'); 
-statistics_pos = fopen([output_folder filesep 'img_statistics_pos.txt'], 'wt');
+if cs == 1 % hsv
+    cs_tag = 'hsv_'; % colorspace tag
+elseif cs == 2 % lab
+    cs_tag = 'cielab_';
+end
+
+statistics_pre_pos = fopen([output_folder_diagnostics filesep strcat(cs_tag, 'img_stats_pre_pos.txt')], 'wt');
 
 % Setting data structure
-fprintf(statistics_pre, 'Img\tMean_hsv_pre\tSD_hsv_pre\n'); 
-fprintf(statistics_pos, 'Img\tMean_hsv_pos\tSD_hsv_pos\n'); 
+fprintf(statistics_pre_pos, 'Img_pre\tMean_pre\tSD_pre\tImg_pos\tMean_pos\tSD_pos\n');
 
 % Setting initial Img
 img1 = [];
 
 % Setting initital number of hsv and lab imgs
-n_hsv_pre = 0;
-n_hsv_pos = 0;
+n_pre = 0;
+n_pos = 0;
 
 % Setting initial mean and standard deviation
-M_v_pre = 0;
-Sd_v_pre = 0;
+M_pre = 0;
+Sd_pre = 0;
 
-M_v_pos = 0;
-Sd_v_pos = 0;
+M_pos = 0;
+Sd_pos = 0;
 
 for i = 1:numim
    % load images
    file_name1 = strcat(input_folder, filesep, src_input(i).name); 
    I1 = imread(file_name1);
-       %figure,imshow(I); % To check if imgs have been loaded properly delete the "%" mark
+       %figure,imshow(I1); % To check if imgs have been loaded properly delete the "%" mark
    
    file_name2 = strcat(output_folder, filesep, src_output(i).name); 
    I2 = imread(file_name2);
-       %figure,imshow(I); % To check if imgs have been loaded properly delete the "%" mark
-    
-   % from rbg to hsv
-   hsv1 = rgb2hsv(I1);
[…]
