+++ toolbox/sfPlot.m
@@ -35,14 +35,24 @@
 % Revision 1, 26/11/2010 (thanks to Diederick C. Niehorster)
 % - Improved speed by using accumarray 
 % - Output argument added; plot is now optional
+% 
+% ------------------------------------------------------------------------
+% SHINE_color toolbox, September 2021, version 0.0.3
+% (c) Rodrigo Dal Ben (dalbenwork@gmail.com)
+%
+% Adapted for diagnostics plot. 
+% Added:
+%   - lum2scale function
+%   - cs & diag input
+% ------------------------------------------------------------------------
 
-function avg = sfPlot(im,qplot)
+function avg = sfPlot(im, qplot, cs, diag)
 
-if nargin < 2
+if nargin < 2 
     qplot = true;
 end
 if ndims(im) == 3
-    im = v2scale(im); % SHINE_color: replaced rgb2gray(im1) for a function that scales hsv Value channel
+    im = lum2scale(im, cs); % SHINE_color: replaced rgb2gray(im1) for a function that scales hsv Value channel
 end
 [xs ys] = size(im);
 fftim = abs(fftshift(fft2(double(im)))).^2;
@@ -58,7 +68,11 @@ end
 avg = accumarray(r(:)+1,fftim(:))./accumarray(r(:)+1,1);
 avg = avg(2:floor(min(xs,ys)/2)+1);
 if qplot
-    figure;loglog(1:floor(min(xs,ys)/2),avg);
-    xlabel('Spatial frequency (cycles/image)');
-    ylabel('Energy');
+    if diag % fig and labs are opened in diag
+        loglog(1:floor(min(xs,ys)/2),avg);
+    else
+        figure;loglog(1:floor(min(xs,ys)/2),avg);
+        xlabel('Spatial frequency (cycles/image)');
+        ylabel('Energy');
+    end
 end

[…]
