+++ toolbox/SHINE_color_INPUT/samples/readme_references.txt
@@ -1,10 +1,16 @@
 If you are using the SHINE_color toolbox, please cite the original SHINE toolbox:
 
-Willenbockel, V., Sadr, J., Fiset, D., Horne, G. O., Gosselin, F., & Tanaka, J. W. (2010). Controlling low-level image properties: The SHINE toolbox. Behavior Research Methods, 42(3), 671–684. http://doi.org/10.3758/BRM.42.3.671
+Willenbockel, V., Sadr, J., Fiset, D., Horne, G. O., Gosselin, F., & Tanaka, J. W. (2010). Controlling low-level image properties: The SHINE toolbox. Behavior Research Methods, 42(3), 671√ê684. http://doi.org/10.3758/BRM.42.3.671
 
 and the SHINE_color OSF project:
 
-Dal Ben, R. (2019). Lum_fun: A set of tools to control luminance of colored images (Version 0.1). [Computer program]. doi: 10.17605/OSF.IO/AUZJY, retrieved from https://osf.io/auzjy/
+Dal Ben, R. (2021). SHINE_color and Lum_fun: A set of tools to control luminance of colorful images (Version 0.0.3). [Computer program]. doi: 10.17605/OSF.IO/AUZJY, retrieved from https://osf.io/auzjy/
 
 
-If you are using any of the NOUN images, please cite the database:Horst, J. S., & Hout, M. C. (2016). The Novel Object and Unusual Name (NOUN) Database: A collection of novel images for use in experimental research. Behavior Research Methods, 48(4), 1393–1409. http://doi.org/10.3758/s13428-015-0647-3
+If you are using any of the NOUN images, please cite the database:
+
+Horst, J. S., & Hout, M. C. (2016). The Novel Object and Unusual Name (NOUN) Database: A collection of novel images for use in experimental research. Behavior Research Methods, 48(4), 1393√ê1409.√ähttp://doi.org/10.3758/s13428-015-0647-3
+
+
+
+
