+++ .gitignore.txt
@@ -1,2 +1,0 @@
-# ignore "00_untracked" folder
-00_untracked/
