+++ paper/paper.md
@@ -15,54 +15,70 @@ date: 21 June 2021
 bibliography: paper.bib
 ---
 
+<!--
+- Detailed flowchart: https://osf.io/uxqtv/
+- add flowchart - DONE
+- Upd table with functions; - DONE
+- Upd text to reflect v0.0.3 upds; - DONE
+- Add Figure showing how the images change (like SHINE), w/ diag_plot; - DONE
+- Fix reviewers suggestions; - DONE
+- delete PDF from repo; - DONE
+- clean toolbox folder;
+- commit changes to diag_plot;
+-->
+
+
 # Summary
 
-Pupil dilation responses are versatile measures that can be used to study an array of complex psychological phenomena. However, controlling stimuli properties is an essential step to minimize experimental confounds that can result in unwanted pupil dilation. Here we describe an adapted MATLAB toolbox for controlling properties of colorful images.
+Some experiments require precise control of visual stimuli properties to avoid experimental confounds (e.g., pupillometry studies). Here we describe an adapted MATLAB toolbox for controlling low-level properties of colorful images.
 
 # Statement of need
 
-Pupil dilation responses can be used to investigate an array of cognitive abilities across the lifespan [@Hepach2016; @Sirois2014; @Zhang2019; @Zhao2019a]. Whereas it is a versatile measure of high-level abilities, pupil dilation can be greatly affected by low-level properties of stimuli and experimental setting such as luminance of visual stimuli and experimental room [e.g., @Hepach2016; @Tsukahara2020].
+One powerful way to control low-level properties of experimental stimuli is to use the SHINE toolbox for MATLAB [@Willenbockel2010]. This toolbox contains a set of functions that allows users to precisely specify luminance and contrast, histogram, and Fourier amplitude spectra of visual stimuli. These parametric manipulations minimize potential low-level confounds when investigating higher-level processes (e.g., cognitive effort, recognition). However, SHINE only works with greyscale images. Whereas this serves well to many research purposes [e.g., @Lawson2017; @Rodger2015], other research goals might benefit from colorful images [e.g., @Cheng2019; @Hepach2016; @Zhang2019]. Here, we describe the SHINE_color, an adaptation of SHINE that allow users to perform all operations from SHINE toolbox on colorful images.
 
-One powerful way to control low-level properties of experimental stimuli is to use the SHINE toolbox for MATLAB [@Willenbockel2010]. This toolbox contains a set of functions that allows users to precisely specify luminance and contrast, histogram, and Fourier amplitude spectra of visual stimuli. These parametric manipulations minimize potential low-level confounds when investigating higher-level processes (e.g., cognitive effort, recognition). However, SHINE only works with greyscale images. Whereas this serves well to many research purposes [e.g., @Lawson2017; @Rodger2015], other research goals might benefit from colorful images [e.g., @Cheng2019; @Hepach2016; @Zhang2019]. Here, we describe the SHINE_color, an adaptation of SHINE that allow users to perform all operations from SHINE toolbox to colorful images.
+SHINE_color might be especially useful for pupillometry studies. Whereas pupil dilation responses can be used to investigate an array of high-level cognitive abilities across the lifespan [@Hepach2016; @Sirois2014; @Zhang2019; @Zhao2019a], it can be greatly affected by low-level properties of stimuli and experimental setting [e.g., @Hepach2016; @Tsukahara2020].
 
 # Implementation
 
[…]
