### 2021-09-24 #3449 @RodDalBen
Hi @takuma929 I have revised the manuscript and the toolbox following your suggestions (i.e., added diagnostic plots, reversed the order of the text in the sections summary and the statement, + minor changes). I will mark your comments as closed, but I will be looking forward for new comments.

Hi @da5nsy, following our discussion [here](https://github.com/RodDalBen/SHINE_color/issues/6), I've added to the toolbox the option to choose between HSV and CIELab color spaces. I also edited the manuscript to reflect this and other updates.
