### 2021-09-24 RodDalBen/SHINE_color#6 @RodDalBen
please, also see [this answer](https://github.com/openjournals/joss-reviews/issues/3449#issuecomment-926722578).
