### 2021-09-24 #3449 @da5nsy
> following our discussion here, I've added to the toolbox the option to choose between HSV and CIELab color spaces. I also edited the manuscript to reflect this and other updates.

Thanks @RodDalBen, and apologies for my lack of reply to your thoughtful and in-depth message on that thread yet. I'll get there soon I hope.
