### 2021-10-04 RodDalBen/SHINE_color#8 @takuma929
Additional comments regarding diagnostic images - JOSS review
> Hi @takuma929 I have revised the manuscript and the toolbox following your suggestions (i.e., added diagnostic plots, reversed the order of the text in the sections summary and the statement, + minor changes). I will mark your comments as closed, but I will be looking forward for new comments.

Hi @RodDalBen thanks for your revision work. I have checked the revised code and manuscript.
My comments were essentially addressed but there are two points that I noticed.

1. Figure 2, on the right side, luminance histograms seems to be completely matched between cat1 and cat2 but slightly different for cat3. Could you check if this really the case?

2. It is great that there is now an option to save diagnostic images. Thanks for implementing this.

However, when I ran the "SHINE_color" with example images "2001-600.jpg", "2002-600.jpg" and "2003-600.jpg" using the default option, I've got two attached images.

In my command window, I've got a message "specMatch successful" and "histMatch successful", but these images show that they don't seem to be matched.. Could you check if luminance histograms and power spectra are properly plotted?

![hsv_histogram_pre_pos](https://user-images.githubusercontent.com/23462512/135851295-b02dfe11-9814-4688-aa2a-31d7b5cc8fda.png)
![hsv_spectrum_pre_pos](https://user-images.githubusercontent.com/23462512/135851297-d6d50602-d646-460d-8542-9500285bd5db.png)
