### 2021-10-04 #3449 @takuma929
> Hi @takuma929 I have revised the manuscript and the toolbox following your suggestions (i.e., added diagnostic plots, reversed the order of the text in the sections summary and the statement, + minor changes). I will mark your comments as closed, but I will be looking forward for new comments.

Thanks for your revision work, @RodDalBen. My comments were essentially addressed.
I have posted some minor additional comments so please check the link below.

https://github.com/RodDalBen/SHINE_color/issues/8#issue-1015117698
