### 2021-10-19 1f7fb1d
v0.0.4: upd lum_calc, diag_plots.
