+++ toolbox/diag_plots.m
@@ -10,12 +10,16 @@
 % Plots saved in SHINE_color_OUTPUT/DIAGNOSTICS.
 %  
 % ------------------------------------------------------------------------
+% SHINE_color toolbox, October 2021, version 0.0.4
+% (c) Rodrigo Dal Ben
+%
+% - Plots are now done directly on the manipulated channels (pre and
+% pos). The previous version re-read & transformed rgb images.
+% ------------------------------------------------------------------------
 
-function diag_plots(input_folder, output_folder, imformat, cs, mode)
+function diag_plots(images_orig, images, imname, cs, mode)
 
-% Set input and output source and name pattern
-src_input = dir(fullfile(strcat(input_folder, filesep, '*.', imformat))); 
-src_output = dir(fullfile(strcat(output_folder, filesep, '*.', imformat)));
+% set input and output dir
 output_folder_diagnostics = fullfile(pwd,'SHINE_color_OUTPUT', 'DIAGNOSTICS');
 
 % set mode
@@ -38,7 +42,7 @@ elseif md == 6 || md == 8
     md2 = [2, 4]; % histogram + spectrum
 end
                 
-% set cs tag
+% set color space tag
 if cs == 1 % hsv
     cs_tag = 'hsv';
 elseif cs == 2 % lab
@@ -46,25 +50,10 @@ elseif cs == 2 % lab
 end
 
 % set img indices for loops
-numim = length(src_input);
+numim = length(images_orig);
 nummd = length(md2);
 j = 1:2:numim*2;
 k = 2:2:numim*2;
-
-% read images
-for z = 1:numim
-    file_name1 = strcat(input_folder, filesep, src_input(z).name); 
-    file_name2 = strcat(output_folder, filesep, src_output(z).name);
-    
-    % rgb images
-    img1{z} = imread(file_name1);
-    img2{z} = imread(file_name2);
-    
-    % scaled luminance
-    lum_img1{z} = lum2scale(img1{z}, cs);
-    lum_img2{z} = lum2scale(img2{z}, cs);
-    
-end
 
 % set number of figures
 if md ~= 1
@@ -82,18 +71,20 @@ if md ~= 1
             if md == 2
                 plot_name = '_histogram_pre_pos';
                 % for a bar plot with thicker bars, uncomment next 2 lines and comment 'imhist'
-                %[counts, grayLevels] = imhist(lum_img1{i}, 256);
+                %[counts, grayLevels] = imhist(images_orig{i}, 256);
                 %bar(grayLevels, counts, 'BarWidth', 5),
-                imhist(lum_img1{i});
-                title(src_input(i).name, 'FontSize', 8)
+                imhist(images_orig{i});
+                title(imname{i}, 'FontSize', 8);
+            
             elseif md == 3
                 plot_name = '_spatial_freq_pre_pos';
-                sfPlot(img1{i}, true, cs, true); 
-                title(src_input(i).name, 'FontSize', 8)
+                sfPlot(images_orig{i}, true, cs, true); 
+                title(imname{i}, 'FontSize', 8)
+            
             elseif md == 4
                 plot_name = '_spectrum_pre_pos';
-                spectrumPlot(img1{i}, true, cs, true); 
-                title(src_input(i).name, 'FontSize', 8)                     
+                spectrumPlot(images_orig{i}, true, cs, true); 
+                title(imname{i}, 'FontSize', 8)                     
             end
             % set axis limits
             axis(sp(i), 'tight');
@@ -101,11 +92,10 @@ if md ~= 1
             for l = 2:2:numim*2
                 sp(i+1) = subplot(numim, 2, k(i));
                 if md == 2
[…]
