### 2021-10-19 RodDalBen/SHINE_color#8 @RodDalBen
Hi @takuma929,

Thank you for your careful review. I couldn't replicate the error with images "2001-600.jpg", "2002-600.jpg", and "2003-600.jpg". However, your comment prompted me to improve the `lum_calc` and `diag_plots` functions. In the previous version, these functions re-read the RGB images from `SHINE_color_input` and `SHINE_color_output` folders, transformed it to either HSV or CIELab, and finally computed the summary statistics (`lum_calc`) or created the diagnostic plots (`diag_plots`). As you carefully noticed, these transformations introduced some noise and inaccuracy in both the stats summary and the plots. Also, they took some time to run. In the new version, both functions are feed directly with luminance channels that are handled by `SHINE` functions (e.g., histogram matching) before and after the manipulation, and always before transforming the HSV or CIELab back to RGB. This modification made the functions more accurate and faster.

I updated Figure 2 in the manuscript with the histogram matching for the three cats (I also added the cats pictures in the sample folder). Their `post` histograms are exactly the same. So are the histograms from "2001-600.jpg", "2002-600.jpg", and "2003-600.jpg", after using the `Default` option, please see below.

One last comment, when using the `Default` option, or any other combination of functions, [Willenbockel et al. (2010)](https://link.springer.com/article/10.3758%2FBRM.42.3.671) recommend using multiple iterations to more closely match the manipulated properties.

I will close this issue for now, but I am looking forward for your next comments and suggestions. Thank you, again for your careful reading!

![hsv_histogram_pre_pos](https://user-images.githubusercontent.com/41629567/138005632-39448de7-a876-45e8-bfe8-27239070744e.png)
![hsv_spectrum_pre_pos](https://user-images.githubusercontent.com/41629567/138005711-03d7c715-7ab3-41ec-b97c-2e08f7bcb239.png)
