### 2021-10-19 #3449 @RodDalBen
@takuma929, thank you for your comments and suggestions.  I have posted my response [here](https://github.com/RodDalBen/SHINE_color/issues/8#issuecomment-947192106).
