### 2021-12-08 #3449 @fredericgosselin
I realize that I’m arriving quite late in this rather long review process. I’m not sure why this is so but, given that I've been asked and that I’ve accepted the invitation, I’ll do my job thoroughly at the risk of setting things back…

To start with, I have three spontaneous reactions or comments to make about this submission:

1)	In the paper, SHINE_color is presented in the context of pupil dilatation research, which, I presume, is the author’s main area of research. This drastically underplays the importance of controlling for low-level image properties. Doing so is relevant to many, if not most, experiments in psychology and cognitive neuroscience as explained in Willenbockel et al. (2010), the original SHINE paper, and as confirmed by the citation count of this original SHINE paper (818 citations according to Google Scholar).

2)	Dal Ben should add a discussion of color monitor calibration in their paper (for example, see Metha, Vingrys & Badcock, 1993) similar to the section "SHINE and Monitor Calibration" about luminance calibration in Willenbockel et al. (2010).

3)	Is SHINE_color a significant enough contribution to warrant a new toolbox and a JOSS publication? In the JOSS documentation, it is stated that “Submissions under 1000 Lines Of Code (LOC) will usually be flagged, those under 300 LOC will be desk rejected.” I’m assuming that it really refers to LOC contributed by the author. In other words, here, LOC not included in the original SHINE toolbox. My estimation is that there are less than 400 new LOC in the SHINE_color toolbox (less than 10% of all LOC in SHINE_color!). If these few additional LOC were implementing a novel algorithm, I would consider these new LOC to be enough to warrant publication in JOSS (perhaps not a new toolbox though). But the "recipe" used by Dal Ben to adapt the original SHINE toolbox to color images is essentially the one given in the section “Applying SHINE to Color Images” in Willenbockel et al. (2010), and the remaining new LOC mostly manage the color dimension of color images.
