### 2021-12-13 #3449 @ajstewartlang
Just checking in with you @RodDalBen to get your thoughts/responses re. the comments made above by @fredericgosselin
