### 2021-12-13 77c5e20
edit intro, emphasis on SHINE, monitor calibration
