+++ paper/paper.md
@@ -17,19 +17,17 @@ bibliography: paper.bib
 
 # Summary
 
-Some experiments require precise control of visual stimuli properties to avoid experimental confounds (e.g., pupillometry studies). Here we describe an adapted MATLAB toolbox for controlling low-level properties of colorful images.
+Many experiments in Psychology and Cognitive Neuroscience require precise control of visual stimuli properties, either to precise manipulate variables of interest or to avoid experimental confounds. One way to control low-level properties of images is to use the SHINE toolbox [@Willenbockel2010], which has been used and cited hundreds of times across a wide range of research topics. Here we describe an adaptation of the SHINE toolbox for controlling low-level properties of colorful images, dubbed SHINE_color.
 
 # Statement of need
 
-One powerful way to control low-level properties of experimental stimuli is to use the SHINE toolbox for MATLAB [@Willenbockel2010]. This toolbox contains a set of functions that allows users to precisely specify luminance and contrast, histogram, and Fourier amplitude spectra of visual stimuli. These parametric manipulations minimize potential low-level confounds when investigating higher-level processes (e.g., cognitive effort, recognition). However, SHINE only works with greyscale images. Whereas this serves well to many research purposes [e.g., @Lawson2017; @Rodger2015], other research goals might benefit from colorful images [e.g., @Cheng2019; @Hepach2016; @Zhang2019]. Here, we describe the SHINE_color, an adaptation of SHINE that allow users to perform all operations from SHINE toolbox on colorful images.
-
-SHINE_color might be especially useful for pupillometry studies. Whereas pupil dilation responses can be used to investigate an array of high-level cognitive abilities across the lifespan [@Hepach2016; @Sirois2014; @Zhang2019; @Zhao2019a], it can be greatly affected by low-level properties of stimuli and experimental setting [e.g., @Hepach2016; @Tsukahara2020].
+One powerful way to control low-level properties of visual experimental stimuli is to use the SHINE toolbox for MATLAB [@Willenbockel2010]. This toolbox contains a set of functions that allows users to precisely specify luminance and contrast, histogram, and Fourier amplitude spectra of visual stimuli. These parametric manipulations minimize potential low-level confounds when investigating higher-level processes (e.g., cognitive effort, recognition). However, SHINE only works with greyscale images. Whereas this serves well to many research purposes [e.g., @Lawson2017; @Rodger2015], other research goals might benefit from colorful images [e.g., @Cheng2019; @Hepach2016; @Zhang2019]. Here, we describe the SHINE_color, an adaptation of SHINE that allow users to perform all operations from SHINE toolbox on colorful images. Such adaptation can be useful for a wide array of research topics that rely on the precise low-level properties of colorful visual stimuli, such as developmental pupillometry studies [@Hepach2016; @Sirois2014; @Zhang2019; @Zhao2019a; @Tsukahara2020].
 
 # Implementation
 
[…]
