### 2021-12-13 #3449 @RodDalBen
Sorry for the late reply, @ajstewartlang, it has been a busy week.

I really appreciate your careful review @fredericgosselin. Here are my responses, following the order of your comments:

1. Indeed, SHINE_color was develop during a pupillometry project as an attempt to avoid matching luminance by hand using Photoshop (a common practice at my former department). I agree that controlling low-level properties of images is relevant for many experiments in Psychology and Cognitive Neuroscience. I have edited the Summary and Statement of Need accordingly (see a new version of the manuscript below).

2. Thank you for highlighting this important step. The manuscript now directs readers to the Monitor Calibration section from the SHINE paper (Willenbockel et al., 2010) and also reminds readers of the importance of the overall lightning conditions on their setup. In addition, I added a message on the script reminding users to calibrate their monitors. Finally, because SHINE_color is an adaptation of SHINE, I initially assumed that readers would ultimately refer to SHINE paper for detailed explanations of the functions and their assumptions. Nonetheless, now the manuscript explicitly encourage readers to refer to Willenbockel et al. (2010).

3. I am not sure if this comment is addressed to me or to the editor, @ajstewartlang. In any case, here are a few thoughts. I agree that SHINE_color **fully** depends on SHINE algorithm, this has always been the intention. The "recipe" is indeed available on Willenbockel et al. (2010), so nothing original in this sense. But, to my knowledge, SHINE_color is the first implementation of such "recipe" in a reproducible and open manner. It comes with open repositories (GitHub and OSF) that foster further collaborative developments. SHINE_color also comes with new features such as video manipulation, diagnostic statistics, and diagnostic plots. Finally, since its' preprint has been [posted on PsyArXiv](https://psyarxiv.com/fec6x/), 5 months ago, it has been downloaded more than 2k times. This, combine with email exchanges with researchers that are using it to study various topics (e.g., bleaching extent of giant clams in the Philippines, archeology in Germany,  emotional development in US, pupillometry in Canada, language development in Australia, Neuroscience in Brazil), adds to the value of SHINE_color and might indicate its' likelihood of being cited and of increasing the visibility and citation of SHINE.
