### 2021-12-17 #3449 @ajstewartlang
I've been through the submission myself - thanks @fredericgosselin for flagging up that much of this submission is code from the original SHINE toolbox and that the novel lines of code are much fewer than we would expect for a JOSS contribution. For this reason, I recommend a reject for this submission.
