### 2021-12-17 #3449 @RodDalBen
Hi @ajstewartlang , thank you for reviewing the submission and making an editorial decision. Just to keep things crystal clear, SHINE_color was always presented as an adaptation of SHINE. Adapted and additional functions were always clearly indicated in Table 1 and all new lines of code (adaptations, new functions, implementations) were always clearly commented on scripts. Finally, I would like to thank @takuma929 and @fredericgosselin, once more, for their careful reviewing. Cheers
