### 2021-12-17 RodDalBen/SHINE_color#6 @RodDalBen
From v0.0.3, SHINE_color supports HSV and CIELab colorspaces. I will close this for now, but leave a comment if you are reading this and would like to work with another colorspace.
